create drush runtime config file, config sync specific module filters
Use drushrc file to cut down typing out lengthy and repetitive command line options in the Drush commands you use and to avoid mistakes.

At the bottom of the file we've added a command specific execution options for the config-import and config-export commands, to ensure that supplied modules are neither enabled nor disabled on import and also that drush does not write over the current export file's state of these module on export.

So we can have devel enabled locally but it should never be enabled on production.

TODO: configure sync-enable so that we can also copy the db down and not have to re-enable devel.