Consolidate google js for search form and search results
Since we are showing both a search form and results on the same page, we can share common google js and draw to both container classes in the callback function.

Note: consolidating the js and having the URL param parsing functionality in the same block as a search form solved the issue of reflecting the search query in the search input.  Win.